% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the scConvertShiny Shiny application}
\usage{
run_app(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Open the app in the default browser.}

\item{...}{Passed on to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
The value returned by \code{\link[shiny:runApp]{shiny::runApp()}}, invisibly.
}
\description{
Run the scConvertShiny Shiny application
}
